fix lỗi object đi qua trái nhanh gấp đôi đi qua phải do sai số khi chuyển từ float sang int

+++ NewSuperMarioBrosPC/object.cpp
@@ -39,8 +39,8 @@ void Object::setSprite(CSprite* sprite){
 	Object::mSprite = sprite;
 }
 void Object::update(int t){
-	x += vx * t;
-	y += vy * t;
+	x += (int)(vx * t);
+	y += (int)(vy * t);
 	vy += ay*t;
 	vx += ax*t;
 	ay -= CMarioGame::GRAVITY_VELOCOTY;
